Always create a new connection to LDAP
There are multiple ways of checking if the LDAP connection has been
closed, but the ldap3 library underpinning BituLDAP cannot check if the
server has dropped the connection, until it attempts to preform an
action. This leads to the stored singleton losing its connection
silently. Instead always request a new connection, because performing an
action.

Change-Id: I23a498b622f175ef1bed537ec77988cf83bced52